checkpoint: fix outgoing messages

+++ src/msg.h
@@ -9,11 +9,11 @@
 
 void write_sid(char *, const nih_t);
 
-u64 msg_q_gp(char *restrict buf, const nih_t nid, const nih_t ih, u16 tok);
-u64 msg_q_fn(char *restrict buf, const pnode_t dest, const nih_t target);
-u64 msg_q_pg(char *restrict buf, const nih_t nid);
-u64 msg_r_fn(char *restrict buf, const parsed_msg *, const pnode_t);
-u64 msg_r_gp(char *restrict buf, const parsed_msg *, const pnode_t);
-u64 msg_r_pg(char *restrict buf, const parsed_msg *);
+u32 msg_q_gp(char *restrict buf, const nih_t nid, const nih_t ih, u16 tok);
+u32 msg_q_fn(char *restrict buf, const pnode_t dest, const nih_t target);
+u32 msg_q_pg(char *restrict buf, const nih_t nid);
+u32 msg_r_fn(char *restrict buf, const parsed_msg *, const pnode_t);
+u32 msg_r_gp(char *restrict buf, const parsed_msg *, const pnode_t);
+u32 msg_r_pg(char *restrict buf, const parsed_msg *);
 
 #endif // DHT_MSG_H
